// Frontend/src/app/processos/models/parametrizar-distribuicao-processos/escritorio-distribuicao.model.ts
export class EscritorioDistribuicaoModel {
    codParamDistribEscrit: number;
    codParamDistribuicao: number;
    codProfissional: number;
    nomProfissional: string;
    codSolicitante: number;
    nomSolicitante: string;
    datVigenciaInicial: Date;
    datVigenciaFinal: Date;
    porcentagemProcessos: string;
    prioridade: number;
}